Added Current Progress
I have been working on this Endless Roller game for a few weeks so far.
I have managed to create a one click endless runner game, with a
randomly generating track. I am currently using a ball in place of the
character, and I am exprermenting with a 2D camera